move TestMnemonic -> testutil
