+++ testutil/known_values.go
@@ -1,4 +1,4 @@
-package tests
+package testutil
 
 const (
 	// Tests expect a ledger device initialized to the following mnemonic
